beginnning to move config files to plugins
